+++ janitor/functions.py
@@ -3259,7 +3259,9 @@ def take_first(
 
 
 @pf.register_dataframe_method
-def shuffle(df: pd.DataFrame, random_state=None) -> pd.DataFrame:
+def shuffle(
+    df: pd.DataFrame, random_state=None, reset_index=True
+) -> pd.DataFrame:
     """
     Shuffle the rows of the DataFrame.
 
@@ -3277,7 +3279,10 @@ def shuffle(df: pd.DataFrame, random_state=None) -> pd.DataFrame:
     :param df: A pandas DataFrame
     :param random_state: (optional) A seed for the random number generator.
     """
-    return df.sample(frac=1, random_state=random_state)
+    result = df.sample(frac=1, random_state=random_state)
+    if reset_index:
+        result = result.reset_index(drop=True)
+    return result
 
 
 @pf.register_dataframe_method
@@ -3397,4 +3402,86 @@ def flag_nulls(
 
     df = df.copy()
     df[column_name] = null_array.astype(int)
+    return df
+
+
+@pf.register_dataframe_method
+def count_cumulative_unique(
+    df: pd.DataFrame,
+    column_name: str,
+    dest_column_name: str,
+    case_sensitive: bool = True,
+) -> pd.DataFrame:
+    """
+    Generates a running total of cumulative unique values in a given column.
+
+    Functional usage example:
+
+    .. code-block:: python
+
+        import pandas as pd
+        import janitor as jn
+
+        df = pd.DataFrame(...)
+
+        df = jn.functions.count_cumulative_unique(
+            df=df,
+            column_name='animals',
+            dest_column_name='animals_unique_count',
+            case_sensitive=True
+        )
+
+    Method chaining usage example:
+
+    .. code-block:: python
+
+        import pandas as pd
+        import janitor.functions
+
+        df = pd.DataFrame(...)
+
+        df = df.count_cumulative_unique(
+            column_name='animals',
+            dest_column_name='animals_unique_count',
+            case_sensitive=True
+        )
+
+    A new column will be created containing a running
+    count of unique values in the specified column.
+    If `case_sensitive` is `True`, then the case of
+    any letters will matter (i.e., 'a' != 'A');
+    othewise, the case of any letters will not matter.
+
+    This method mutates the original DataFrame.
+
+    :param df: A pandas dataframe.
+    :param column_name: Name of the column containing
+        values from which a running count of unique values
+        will be created.
+    :param dest_column_name: The name of the new column containing the
+        cumulative count of unique values that will be created.
+    :param case_sensitive: Whether or not uppercase and lowercase letters
+        will be considered equal (e.g., 'A' != 'a' if `True`).
+
+    :returns: A pandas DataFrame with a new column containing a cumulative
+        count of unique values from another column.
+    """
+
+    if not case_sensitive:
+        # Make it so that the the same uppercase and lowercase
+        # letter are treated as one unique value
+        df[column_name] = df[column_name].astype(str).map(str.lower)
+
+    df[dest_column_name] = (
+        (
+            df[[column_name]]
+            .drop_duplicates()
+            .assign(dummyabcxyz=1)
+            .dummyabcxyz.cumsum()
+        )
+        .reindex(df.index)
+        .ffill()
+        .astype(int)
+    )
+
     return df